Create TrailingZeroesFactorial1
We find trailing zeroes in the factorial of a number entered by user.
We can find factorial of that number and see trailing zeroes in it but that would be time consuming and tough for bigger inputs.
So, here we used the concept that zeroes are made of 10(2X5).
And infact 5s will be less than 2 so we can use all the 5s to do that.
Trailing Zeroes Count = (n/5)+(n/25)+(n/125)+.........
For Example : in 10! = 10x9x8x7x6x5x4x3x2x1
10/5=2 which is our answer 
as there is a pair of 2 and 5 as well as a 10 which is also basically a pair of 2 and 5.